qt5-doc: Switch to a pre-built documentation tarball.
The existing port was not working with Qt 5.9.

Instead of trying to fix it, switch to what Gentoo does and fetch a
tarball that already contains all documentation so that we do not have to
build anything at all.

The tarball's name and location in download.qt.io look a bit weird, but it
seems to work fine.