Die
Currently both mDNSResponder and avahi-libdns provide the same
library file and therefore are marked to conflict with each other.

However, as only the library is tracked for PORTS runtime dependency
management, it is not possible to distinguish which port actually
provided it at this point.

In addition to the library, add a build time dependency on some
header files, which are exclusive to each of the two ports.
Doing so will have the build die with a PORTS conflict rather than
silently ignoring a user choice.