bsd.qt.mk: Specify the full path in QMAKESPEC for Qt4.
This reverts part of commit 8129829 ("Link in qmake to the
CONFIGURE_WRKSRC/qmake/ for base ports").

Specifying only the name of the mkspec doesn't work in Qt 4, mostly because
we exclude mkspecs/ when extracting the tarball (e.g. in devel/qt4-moc). It
is easier to just preserve compatibility for Qt4-based ports and avoid
changing the Qt4 ports themselves.

Based on a patch by tcberner.