qt5-core: Restore most CONFIGURE_ARGS arguments.
Most arguments were dropped in commit 82de3e831 ("Some stuff for qt5-core")
without any further explanation, and since restoring them does not cause any
build issues it is better to reduce the diff to master.

A few options were not re-added because they no longer exist in the
configure script. "-no-feature-concurrent" was not re-added to avoid having
QT_NO_CONCURRENT in qconfig.h and "concurrent" in the disabled libraries in
the mkspecs.