bsd.qt.mk: Revert to always setting QMAKESPEC to a full path.
This reverts a POLA violation, and Qt5's qmake continues working with full
paths.

It's only the configuration script in qtbase that needs an mkspec name
instead of a full path, so we can narrow our adjustment to that part.

This fixes net-im/telegram-desktop's build, as it does a sed subtitution
that needs a full path for the gyp part of its build system:

    /usr/local/include/qt5/QtCore/5.9.3/QtCore/private/qthread_p.h:56:10: fatal error: 'qplatformdefs.h' file not found