Link kdevelop's used clang with the ports Makefile
Don't rely on cmake to find the "right" clang installation, but
tell it where it's installed instead.

Enabling the cmake policy is unneccessary right now because
kdevelop provides it's own findLLVM.cmake, which ignores the
policy, but this could change at some point in the future.

While here, introduce the LLVM_VERSION make variable and have
it default to 60, which is the version that was used by cmake
on stable/11 before (the LIBDEPENDS in the makefile was not
communicated to cmake before this commit).

Eventually LLVM_VERSION shouldn't be hard-coded here,
but defaulting to MESA_LLVM_VER, which is installed
anyways due to x11 depending on mesa-dri.