Don't rely on cmake to find the "right" clang installation
... but tell it where it's installed instead.

Enabling the cmake policy is unneccessary right now because
kdevelop provides it's own findLLVM.cmake, that ignores the
policy, but this could change at some point in the future.

While here, introduce the LLVM_VERSION make variable and have
it default to 60, which is the version that was used by cmake
on stable/11 before (the LIBDEPENDS in the makefile was not
communicated to cmake before this commit).

Eventually LLVM_VERSION shouldn't be hard-coded here,
but defaulting to MESA_LLVM_VER, which is installed
anyways because x11 depends on it.

This commit has no effect on "default setups", i.e. those
that don't define LLVM_VERSION to ports and don't define
LLVM_ROOT as an environment variable either.
As such nondefault setups probably exist, bump PORTREVISION.