This commit should enable the build with debug symbols unconditionally.
It's rather experimental and has been tested so far with the ALSA only.

Note: It requires a huge amount of memory - on a system with 32GB of RAM there
      should be at least 20GB of swap space to get a successful build.

The *.debug files were removed from the pkg-plist just as a quick fix. The original
pkg-plist is also included as pkg-plist.bak for reference.

After a successful build there is a shared lib with a size of nearly 4GB ram.