Revised the sets of patches since Chromium 71.x has been landed into the ports.
There is a small difference between the workflows of www/qt5-webengine and
www/chromium but this difference has a huge impact of the outcome:

www/qt5-webengine defined in its BUILDCONFIG for the condition (os == "bsd")
two variables:

* is_bsd = true AND is_linux = false

In all the BUILD.gn files one then had to add support for *BSD systems by
using (is_linux || is_bsd) to enable features/sources for *BSD.

On the other hand www/chromium sets in its BUILDCONFIG for the condition
(os == "bsd") the variables as follows:

* is_bsd = true AND is_linux = true.

This dramatically reduces the modification of the BUILD.gn files and one has
to exclude features/sources for *BSD by using (is_linux && !is_bsd).

Some changes seem to be tautological and most of those are pure cosmetical
to get in sync with the latest patches for chromium. Many patches have
been removed because they are meant for modules that www/qt5-webengine won't
compile/use.

Another interesting fact is that the debug build requires now less memory
than before. Maybe it will now be possible to find the last piece to make
www/qt5-webengine more stable and ready for the community.

www/qt5-webengine seems to be a bit more stable since the last commit. It
was possible to play multimedia content in 4-5 tabs before a SIGTRAP was
caught.

Between Chromium 69.x and Chromium 71.x there are again some changes in the
handling of threads and processes. The file base/process/process_posix.cc has
been removed from the patch set because it crashes www/qt5-webengine pretty
fast.

TL;DR; : This is a checkpoint release, build strategy was changed and many
patches were cleaned up or removed. Debug builds requires less memory and
the stability is a bit better than before but not yet optimal.