www/qt5-webengine: Update to 5.15.0-rc (preliminary)
The rc uses Chromium 80.0.3987.163 and no patches needs to be added/removed.

Also rebase two patches to clean-up the output (= "Line XX not found") of
"make patch"

Builds fine on 11.3-, 12.1-RELEASE, 13.0-CURRENT@r359374 amd64 + i386 with
ALSA/PULSEAUDIO/SNDIO and with/-out DEBUG options set. On i386 there's the
known issue with memory exhaustions with DEBUG set to on.

TODO:
- Test build for other platforms (aarch64, ppc64, etc.)
- Runtime tests (not done yet) /!\