devel/qt6: Repurpose 'makesum' target
The default 'makesum' target in the metaport is useless. Redefine it to
generate the distinfo files for all of the Qt6 components. This is part
of an effort to simplify Qt updates for myself and any future
maintainers and helps prevent new components from being forgotten in
bespoke update scripts.