net/wifi-firmware-kmod: generalize iwlwifi specific tunable writing
Currently iwlwifi(4) is the only LinuxKPI based wireless driver able
to enable 11n/11ac.  bz has added tunables to rtw88 and rtw89 in order
to allow them to gradually enable more modern standards as well.

Generalize the ports framework to not be iwlwifi-specifc anymore
in order to prepare for Realtek and others to gain 11n and 11ac
support.

At some point in time, along with a kernel change, we will then flip
the defaults and make this a "disable" rather than an enable option.

Sponsored by:	The FreeBSD Foundation
Reviewers:	jrm
Differential Revision: https://reviews.freebsd.org/D53789