devel/lua-language-server: Resurrect port
Once upon a time, I made a lua-language-server port. It was
basically one giant hack, and it sortof worked, if you stood
on one foot and didn't breathe. Eventually it became too big
of a burden and I deprecated it, and it was subsequently
removed back in 2022.

Dave Marker took the old port, worked with markj to fix some
longstanding bugs in an upstream dependency of a dependency,
and got the port working in a far more stable state than I'd
ever gotten it.

It's worth noting that the build is remarkably complex. LuaLS
is built by luamake (with its baffling DSL) which is, in turn,
built using raw ninja which is.... awful to work with.

It's also worth noting that Dave took this on as his first
port.

So, here comes--once again--the lua-language-server (aka LuaLS),
the most popular and featureful LSP server for Lua.

Differential Revision:	https://reviews.freebsd.org/D54055