www/angie: Use upstream version numbers for modules
Angie modules currently inherit the version number of the angie version
they are built against. This makes it hard and opaque to track their
actual version.
The mechanism used in the Makefile also had the unwanted side-effect of
PORTREVISION propagating from modules to the master port, which broke
the build (we found out the hard way...)

While here reduce number of of MKDIR calls.

PR:		291490
Co-authored-by:	Vladimir Druzenko <vvd@FreeBSD.org>