net/wifi-firmware-ath1[012]k-kmod: update firmware files
With the removal of the "USE=kmod" and only building packages with
plain firmware files, our version went backwards.  Updating the
firmware will fix this.  For ath1[012]k we can just do that as we
do not yet ship the drivers officially.

Change to a script shipping with ath10k to extract the firmware files
from WHENCE to get them all.  It is questionable if we really want
to have as many flavors or how to ship them in the end, especially
for ath11k and ath12k.

Sponsored by:	The FreeBSD Foundation
Reviewed by:	jrm
Differential Revision: https://reviews.freebsd.org/D54200