net/wifi-firmware-rtw8[89]-kmod: update firmware files
With the removal of the "USE=kmod" and only building packages with
plain firmware files, our version went backwards.  Updating the
firmware will fix this.

Change to a script to extract the firmware files from WHENCE to get
them all.  This also helps us as we no longer have to manually track
older files as they are already included in the list.

Sponsored by:	The FreeBSD Foundation
Reviewed by:	jrm
Differential Revision: https://reviews.freebsd.org/D54202