net/wifi-firmware-kmod: update along LinuxKPI firmware ports to 20251125
We update flavors added with the updated wifi-firmware-*-kmod ports;
the only incompatible change is with mt76 in that we dropped mt792x
in favour of mt7921 and mt7925.  Give we did not ship the driver in
any release yet we can do that.  Some older fwget(8) versions would
be cranky but neither of these releases will ever see mt76 support,
so we are fine.

Given all child ports were updated to the same version (20251125),
reset PORTREVISION to 0 (comment it out) in Makefile.inc as it is
no longer needed. [1]

Pointed out by:	jrm [1]
Sponsored by:	The FreeBSD Foundation
Reviewed by:	jrm
Differential Revision: https://reviews.freebsd.org/D54204