devel/janet-lsp: Compile as a portable "binscript"
Janet can compile to a portable "binscript" or a statically-linked
executable. The latter allows the LSP to work even without Janet
installed, but means that the LSP can be out-of-sync if the user
has a different version of Janet installed (the LSP will be unaware
of that new Janet and its capabilities).

To guard against such problems, switch to the binscript approach,
which compiles to an IR. Janet is a required runtime dep now, but
there's no risk of the LSP getting out-of-sync with the underlying
toolchain.

Differential Revision:	https://reviews.freebsd.org/D54260